add multiclass classfication
